requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for full passlib compatibility

# Google OAuth
google-auth==2.27.0

# OpenAI
openai==1.10.0

# Environment
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Development
ruff==0.1.14
black==24.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
